Cleaned up some duplication
